requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.24.0
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.7.0
openai-whisper>=20231117
websockets>=12.0
pydantic-settings>=2.1.0
pyannote.audio>=3.1.1
soundfile>=0.12.1
librosa>=0.10.1
scipy>=1.11.0
huggingface_hub